+++ apps/backend/src/route-handlers/crud-handler.tsx
@@ -1,6 +1,6 @@
 import "../polyfills";
 
-import { Tenancy, getSoleTenancyFromProject } from "@/lib/tenancies";
+import { Tenancy, getSoleTenancyFromProjectBranch, } from "@/lib/tenancies";
 import { traceSpan } from "@/utils/telemetry";
 import { CrudSchema, CrudTypeOf, CrudlOperation } from "@stackframe/stack-shared/dist/crud";
 import { ProjectsCrud } from "@stackframe/stack-shared/dist/interface/crud/projects";
@@ -76,7 +76,7 @@ type CrudHandlerDirectByAccess<
       user?: UsersCrud["Admin"]["Read"],
       allowedErrorTypes?: (new (...args: any) => any)[],
     }
-    & ({ project: ProjectsCrud["Admin"]["Read"], branchId: string } | { tenancy: Tenancy })
+    & ({ project: Omit<ProjectsCrud["Admin"]["Read"], "config">, branchId: string } | { tenancy: Tenancy })
     & (({} extends yup.InferType<QS> ? {} : never) | { query: yup.InferType<QS> })
     & (L extends "Create" | "List" ? Partial<yup.InferType<PS>> : yup.InferType<PS>)
     & (K extends "Read" | "List" | "Delete" ? {} : (K extends keyof T[A] ? { data: T[A][K] } : "TYPE ERROR: something went wrong here"))
@@ -243,21 +243,25 @@ export function createCrudHandlers<
               async ({ user, project, branchId, tenancy, data, query, allowedErrorTypes, ...params }: yup.InferType<PS> & {
                 query?: yup.InferType<QS>,
                 user?: UsersCrud["Admin"]["Read"],
-                project?: ProjectsCrud["Admin"]["Read"],
+                project?: Omit<ProjectsCrud["Admin"]["Read"], "config">,
                 branchId?: string,
                 tenancy?: Tenancy,
                 data: any,
                 allowedErrorTypes?: (new (...args: any) => any)[],
               }) => {
-                if (tenancy && project) {
-                  throw new StackAssertionError("Must specify either project or tenancy, not both");
-                } else if (tenancy) {
+                if (tenancy) {
+                  if (project || branchId) {
+                    throw new StackAssertionError("Must specify either project and branchId or tenancy, not both");
+                  }
                   project = tenancy.project;
                   branchId = tenancy.branchId;
                 } else if (project) {
-                  tenancy = await getSoleTenancyFromProject(project);
+                  if (!branchId) {
+                    throw new StackAssertionError("Must specify branchId when specifying project");
+                  }
+                  tenancy = await getSoleTenancyFromProjectBranch(project.id, branchId);
                 } else {
-                  throw new StackAssertionError("Must specify either project or tenancy");
+                  throw new StackAssertionError("Must specify either project and branchId or tenancy");
                 }
 
                 try {
@@ -268,9 +272,9 @@ export function createCrudHandlers<
                       data,
                       auth: {
                         user,
-                        project: project ?? throwErr("Project not found in CRUD handler invocation", { project, tenancy }),
-                        branchId: branchId ?? throwErr("Branch ID not found in CRUD handler invocation", { project, branchId }),
-                        tenancy: tenancy ?? throwErr("Tenancy not found in CRUD handler invocation", { project, tenancy }),
+                        project: project ?? throwErr("Project not found in CRUD handler invocation", { project, tenancy, branchId }),
+                        branchId: branchId ?? throwErr("Branch ID not found in CRUD handler invocation", { project, tenancy, branchId }),
+                        tenancy: tenancy ?? throwErr("Tenancy not found in CRUD handler invocation", { project, tenancy, branchId }),
                         type: accessType,
                       },
                     });