+++ apps/e2e/tests/backend/backend-helpers.ts
@@ -3,6 +3,7 @@ import { encodeBase64 } from "@stackframe/stack-shared/dist/utils/bytes";
 import { generateSecureRandomString } from "@stackframe/stack-shared/dist/utils/crypto";
 import { StackAssertionError, throwErr } from "@stackframe/stack-shared/dist/utils/errors";
 import { filterUndefined } from "@stackframe/stack-shared/dist/utils/objects";
+import { wait } from "@stackframe/stack-shared/dist/utils/promises";
 import { nicify } from "@stackframe/stack-shared/dist/utils/strings";
 import * as jose from "jose";
 import { randomUUID } from "node:crypto";
@@ -12,6 +13,7 @@ import { Context, Mailbox, NiceRequestInit, NiceResponse, STACK_BACKEND_BASE_URL
 type BackendContext = {
   readonly projectKeys: ProjectKeys,
   readonly defaultProjectKeys: ProjectKeys,
+  readonly currentBranchId: string | null,
   readonly mailbox: Mailbox,
   readonly userAuth: {
     readonly refreshToken?: string,
@@ -33,6 +35,7 @@ export const backendContext = new Context<BackendContext, Partial<BackendContext
   () => ({
     defaultProjectKeys: InternalProjectKeys,
     projectKeys: InternalProjectKeys,
+    currentBranchId: null,
     mailbox: createMailbox(`default-mailbox--${randomUUID()}${generatedEmailSuffix}`),
     generatedMailboxNamesCount: 0,
     userAuth: null,
@@ -127,6 +130,7 @@ export async function niceBackendFetch(url: string | URL, options?: Omit<NiceReq
         "x-stack-super-secret-admin-key": projectKeys.superSecretAdminKey,
         'x-stack-admin-access-token': projectKeys.adminAccessToken,
       } : {},
+      "x-stack-branch-id": backendContext.value.currentBranchId ?? undefined,
       "x-stack-access-token": userAuth?.accessToken,
       "x-stack-refresh-token": userAuth?.refreshToken,
       ...backendContext.value.ipData ? {
@@ -145,7 +149,7 @@ export async function niceBackendFetch(url: string | URL, options?: Omit<NiceReq
     }),
   });
   if (res.status >= 500 && res.status < 600) {
-    throw new StackAssertionError(`Unexpected internal server error: ${res.status} ${typeof res.body === "string" ? res.body : nicify(res.body)}`);
+    throw new StackAssertionError(`API threw ISE in ${otherOptions.method ?? "GET"} ${url}: ${res.status} ${typeof res.body === "string" ? res.body : nicify(res.body)}`);
   }
   if (res.headers.has("x-stack-known-error")) {
     expect(res.status).toBeGreaterThanOrEqual(400);
@@ -336,11 +340,23 @@ export namespace Auth {
     }
 
     export async function signIn() {
-      const mailbox = backendContext.value.mailbox;
       const sendSignInCodeRes = await sendSignInCode();
+      const signInResult = await signInWithCode(await getSignInCodeFromMailbox());
+      return {
+        ...sendSignInCodeRes,
+        ...signInResult,
+      };
+    }
+
+    export async function getSignInCodeFromMailbox() {
+      const mailbox = backendContext.value.mailbox;
       const messages = await mailbox.fetchMessages();
       const message = messages.findLast((message) => message.subject.includes("Sign in to")) ?? throwErr("Sign-in code message not found");
       const signInCode = message.body?.text.match(/http:\/\/localhost:12345\/some-callback-url\?code=([a-zA-Z0-9]+)/)?.[1] ?? throwErr("Sign-in URL not found");
+      return signInCode;
+    }
+
+    export async function signInWithCode(signInCode: string) {
       const response = await niceBackendFetch("/api/v1/auth/otp/sign-in", {
         method: "POST",
         accessType: "client",
@@ -367,7 +383,6 @@ export namespace Auth {
       });
 
       return {
-        ...sendSignInCodeRes,
         userId: response.body.user_id,
         signInResponse: response,
       };
@@ -555,12 +570,13 @@ export namespace Auth {
 
 
   export namespace OAuth {
-    export async function getAuthorizeQuery() {
+    export async function getAuthorizeQuery(options: { forceBranchId?: string } = {}) {
       const projectKeys = backendContext.value.projectKeys;
       if (projectKeys === "no-project") throw new Error("No project keys found in the backend context");
+      const branchId = options.forceBranchId ?? backendContext.value.currentBranchId;
 
       return {
-        client_id: projectKeys.projectId,
+        client_id: !branchId ? projectKeys.projectId : `${projectKeys.projectId}#${branchId}`,
         client_secret: projectKeys.publishableClientKey ?? throwErr("No publishable client key found in the backend context"),
         redirect_uri: localRedirectUrl,
         scope: "legacy",
@@ -572,14 +588,14 @@ export namespace Auth {
       };
     }
 
-    export async function authorize(options?: { redirectUrl?: string, errorRedirectUrl?: string }) {
+    export async function authorize(options: { redirectUrl?: string, errorRedirectUrl?: string, forceBranchId?: string } = {}) {
       const response = await niceBackendFetch("/api/v1/auth/oauth/authorize/spotify", {
         redirect: "manual",
         query: {
-          ...await Auth.OAuth.getAuthorizeQuery(),
+          ...await Auth.OAuth.getAuthorizeQuery(options),
           ...filterUndefined({
-            redirect_uri: options?.redirectUrl ?? undefined,
-            error_redirect_uri: options?.errorRedirectUrl ?? undefined,
+            redirect_uri: options.redirectUrl ?? undefined,
+            error_redirect_uri: options.errorRedirectUrl ?? undefined,
           }),
         },
       });
@@ -591,10 +607,10 @@ export namespace Auth {
       };
     }
 
-    export async function getInnerCallbackUrl(options?: { authorizeResponse: NiceResponse }) {
-      options ??= await Auth.OAuth.authorize();
+    export async function getInnerCallbackUrl(options: { authorizeResponse?: NiceResponse, forceBranchId?: string } = {}) {
+      const authorizeResponse = options.authorizeResponse ?? (await Auth.OAuth.authorize(options)).authorizeResponse;
       const providerPassword = generateSecureRandomString();
-      const authLocation = new URL(options.authorizeResponse.headers.get("location")!);
+      const authLocation = new URL(authorizeResponse.headers.get("location")!);
       const redirectResponse1 = await niceFetch(authLocation, {
         redirect: "manual",
       });
@@ -667,15 +683,23 @@ export namespace Auth {
       expect(innerCallbackUrl.origin).toBe("http://localhost:8102");
       expect(innerCallbackUrl.pathname).toBe("/api/v1/auth/oauth/callback/spotify");
       return {
-        ...options,
+        authorizeResponse,
         innerCallbackUrl,
       };
     }
 
-    export async function getAuthorizationCode(options?: { innerCallbackUrl: URL, authorizeResponse: NiceResponse }) {
-      options ??= await Auth.OAuth.getInnerCallbackUrl();
-      const cookie = updateCookiesFromResponse("", options.authorizeResponse);
-      const response = await niceBackendFetch(options.innerCallbackUrl.toString(), {
+    export async function getAuthorizationCode(options: { innerCallbackUrl?: URL, authorizeResponse?: NiceResponse, forceBranchId?: string } = {}) {
+      let authorizeResponse, innerCallbackUrl;
+      if (options.innerCallbackUrl && options.authorizeResponse) {
+        innerCallbackUrl = options.innerCallbackUrl;
+        authorizeResponse = options.authorizeResponse;
+      } else if (!options.innerCallbackUrl) {
+        ({ authorizeResponse, innerCallbackUrl } = await Auth.OAuth.getInnerCallbackUrl(options));
+      } else {
+        throw new Error("If innerCallbackUrl is provided, authorizeResponse must also be provided");
+      }
+      const cookie = updateCookiesFromResponse("", authorizeResponse!);
+      const response = await niceBackendFetch(innerCallbackUrl.toString(), {
         redirect: "manual",
         headers: {
           cookie,
@@ -701,7 +725,7 @@ export namespace Auth {
       };
     }
 
-    export async function signIn() {
+    export async function signIn(options: { forceBranchId?: string } = {}) {
       const getAuthorizationCodeResult = await Auth.OAuth.getAuthorizationCode();
 
       const projectKeys = backendContext.value.projectKeys;
@@ -996,6 +1020,11 @@ export namespace Project {
       body: {
         display_name: body?.display_name || 'New Project',
         ...body,
+        config: {
+          credential_enabled: true,
+          allow_localhost: true,
+          ...body?.config,
+        },
       },
     });
     expect(response).toMatchObject({
@@ -1098,7 +1127,7 @@ export namespace Team {
     };
   }
 
-  export async function createAndAddCurrent(options: { accessType?: "client" | "server" } = {}, body?: any) {
+  export async function createWithCurrentAsCreator(options: { accessType?: "client" | "server" } = {}, body?: any) {
     return await Team.create({ ...options, addCurrentUser: true }, body);
   }
 
@@ -1182,15 +1211,24 @@ export namespace Team {
 
 export namespace User {
   export function setBackendContextFromUser({ mailbox, accessToken, refreshToken }: {mailbox: Mailbox, accessToken: string, refreshToken: string}) {
-      backendContext.set({
-        mailbox,
-        userAuth: {
-          accessToken,
-          refreshToken,
-        },
-      });
+    backendContext.set({
+      mailbox,
+      userAuth: {
+        accessToken,
+        refreshToken,
+      },
+    });
   }
 
+  export async function getCurrent() {
+    const response = await niceBackendFetch("/api/v1/users/me", {
+      accessType: "client",
+    });
+    expect(response).toMatchObject({
+      status: 200,
+    });
+    return response.body;
+  }
 
   export async function create({ emailAddress }: {emailAddress?: string} = {}) {
     // Create new mailbox
@@ -1266,6 +1304,23 @@ export namespace Webhook {
       svixToken,
       endpointId: createEndpointResponse.body.id
     };
+  }
+
+  export async function findWebhookAttempt(projectId: string, endpointId: string, svixToken: string, fn: (msg: any) => boolean) {
+    // retry many times because Svix sucks and is slow
+    for (let i = 0; i < 20; i++) {
+      const attempts = await Webhook.listWebhookAttempts(projectId, endpointId, svixToken);
+      const filtered = attempts.filter(fn);
+      if (filtered.length === 0) {
+        await wait(500);
+        continue;
+      } else if (filtered.length === 1) {
+        return filtered[0];
+      } else {
+        throw new Error(`Found ${filtered.length} webhook attempts for project ${projectId}, endpoint ${endpointId}`);
+      }
+    }
+    throw new Error(`Webhook attempt not found for project ${projectId}, endpoint ${endpointId}`);
   }
 
   export async function listWebhookAttempts(projectId: string, endpointId: string, svixToken: string) {