+++ packages/stack-shared/tsup.config.ts
@@ -1,0 +1,3 @@
+import createJsLibraryTsupConfig from '../../configs/tsup/js-library';
+
+export default createJsLibraryTsupConfig({ barrelFile: false });