+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/stack-shared
 
+## 2.8.11
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.10
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.9
+
+## 2.8.8
+
+## 2.8.7
+
 ## 2.8.6
 
 ### Patch Changes